backend completed

+++ backend/views/event_view.py
@@ -4,8 +4,8 @@ from flask_jwt_extended import jwt_required
 from flask_jwt_extended import get_jwt_identity
 from models.events_model import EventsModel
 from schemas.event_schema import event_schema, events_schema
-from datetime import datetime
 from utils.db import db
+from error_handler import CustomError
  
 
 class EventResourceList(Resource):
@@ -22,14 +22,15 @@ class EventResourceList(Resource):
     def post(self):
         user = get_jwt_identity()
         data = request.get_json()
-        new_event = EventsModel(name=data['name'],
-        place= data['place'],
-        keeper_id = user['id'],
-        starts_at= datetime.fromisoformat(data["starts_at"]),
-        ends_at= datetime.fromisoformat(data["ends_at"]) ,
-        address = data['address'],
-        category = data['category'],
-        is_face_to_face = data['is_face_to_face'])
+
+        if len(data.items()) == 0:
+            raise CustomError("Needs to provide info for create", status_code=400)
+        
+        new_event = EventsModel(keeper_id = user['id'])
+
+        
+        for i in data:
+            setattr(new_event, i, data[i])
 
         db.session.add(new_event)
         db.session.commit()
@@ -41,25 +42,37 @@ class EventResourceDetail(Resource):
     @jwt_required()
     def get(self, event_id):
         user = get_jwt_identity()
-        event = EventsModel.query.filter_by(id= event_id, keeper_id=user.id).first()
+        event = EventsModel.query.filter_by(id= event_id, keeper_id=user['id']).first()
         return event_schema.dump(event)
 
     @jwt_required()
     def put(self, event_id):
         user = get_jwt_identity()
-        old_event = EventsModel.query.filter_by(id= event_id, keeper_id=user.id).first()
+        old_event = EventsModel.query.filter_by(id= event_id, keeper_id=user['id']).first()
+
+        if not old_event: 
+            raise CustomError("No such event!", status_code=404)
+
         data = request.get_json()
 
+        if len(data.items()) == 0:
+            raise CustomError("Needs to provide update info", status_code=400)
+
         for i in data:
-            old_event[i] = data[i]
+            setattr(old_event, i, data[i])
 
         db.session.add(old_event)
         db.session.commit()
-    
+        db.session.refresh(old_event)
+        return event_schema.dump(old_event)
+
     @jwt_required()
     def delete(self, event_id):
             user = get_jwt_identity()
-            event_to_delete = EventsModel.query.filter_by(id= event_id, keeper_id=user.id).first()
+            event_to_delete = EventsModel.query.filter_by(id= event_id, keeper_id=user['id']).first()
+            if not event_to_delete: 
+                raise CustomError("No such event!", status_code=404)
+            
             db.session.delete(event_to_delete)
             db.session.commit()
             return '', 204
